requirements.txt
google-genai
google-cloud-documentai
python-dotenv
FastAPI